@JulianKnodt 2022-01-15
If you're willing to put the effort in, I think it would still be possible, but it would require more work. I assume in practice it would hit the worst case less often